Another Win64 build fix
TBR=bmeurer@chromium.org
BUG=

Review URL: https://codereview.chromium.org/527253002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23604 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/js-generic-lowering.cc
@@ -323,8 +323,9 @@ void JSGenericLowering::ReplaceWithCompareIC(Node* node, Token::Value token,
     inputs.push_back(NodeProperties::GetEffectInput(node));
     inputs.push_back(NodeProperties::GetControlInput(node));
   }
-  Node* compare = graph()->NewNode(common()->Call(desc_compare), inputs.size(),
-                                   &inputs.front());
+  Node* compare =
+      graph()->NewNode(common()->Call(desc_compare),
+                       static_cast<int>(inputs.size()), &inputs.front());
 
   node->ReplaceInput(0, compare);
   node->ReplaceInput(1, SmiConstant(token));
